\name{SpamClassification-package}
\alias{SpamClassification-package}
\alias{SpamClassification}
\docType{package}
\title{
\packageTitle{SpamClassification}
}
\description{
\packageDescription{SpamClassification}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SpamClassification}
\packageIndices{SpamClassification}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SpamClassification}

Maintainer: \packageMaintainer{SpamClassification}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
